+++ src/Header/Provider.php
@@ -1,0 +1,14 @@
+<?php
+declare(strict_types = 1);
+
+namespace Innmind\Http\Header;
+
+use Innmind\Http\Header;
+
+/**
+ * @psalm-immutable
+ */
+interface Provider
+{
+    public function toHeader(): Header;
+}